Refactor/#16-폴더 구조 변경
